specs: Add a way to query multiple views from a wayland buffer
the EGL_WL_bind_wayland_display_spec now has an extra attribute to
eglCreateImageKHR to select a particular view of a wayland buffer. This can be
used to get a texture of one of the eyes of a stereoscopic buffer. The buffer
number of views can be queried by passing EGL_MULTIVIEW_VIEW_COUNT_EXT to
eglQueryWaylandBufferWL.