Use the magic behaviour of GL_BACK in GLES 1 and 2 as well as 3
In GLES 3 it is not possible to select rendering to the front buffer and
instead selecting GL_BACK has the magic interpretation that it is either the
front buffer on single-buffered configs or the back buffer on double-buffered.
GLES 1 and 2 have no way of selecting the draw buffer at all. In that case we
were initialising the draw buffer to either GL_FRONT or GL_BACK depending on
the context's config and then leaving it at that. When we switch to having
configless contexts we ideally want Mesa to automatically switch between the
front and back buffer whenever a double- or single-buffered surface is bound.
To make this happen we can just allow the magic behaviour from GLES 3 in GLES
1 and 2 as well. It shouldn't matter what the internal value of the draw
buffer is in GLES 1 and 2 because there is no way to query it from the
external API.