Include the border size when determining the rb size from a texture
When attaching a texture to a framebuffer it was previously using the size of
the texture minus the border when determining the renderbuffer size. However
the size of the image buffer includes the border so when used with the swrast
driver it would end up calculating the wrong stride and it would render
garbage.