intel: Don't flush the old context in intelMakeCurrent
It shouldn't be necessary to flush the context within the driver
implementation because the old context is explicitly flushed in
_mesa_make_current which is called a little further on. It is useful to only
have a single place that flushes when switching contexts to make it easier to
later implement the GL_KHR_context_flush_control extension.

The flush in intelMakeCurrent was added in commit 5505865 to implement the
GLX semantics that the context should be flushed when it is released. When the
commit was made there was no flush in _mesa_make_current because it was only
added later in 93102b4c. I think that later commit effectively makes the first
commit redundant.

The later commit was added because it looks like the flush in intelMakeCurrent
is actually not even hit anymore. This is because when a new context is bound
the old context is usually unbound first in MakeContextCurrent and
dri2_make_current. Since commit b4bb668020 the unbind implementations call
_mesa_make_current(NULL) which effectively means that there should almost
never be a current context when intelMakeCurrent is called.

However one caveat with that is that MakeContextCurrent doesn't call the
unbind virtual if the context is still bound in another thread. It seems to me
like this is a bug anyway because the call to _mesa_make_current(NULL) was
added to fix a problem where the bind implementation can incorrectly think the
context is already bound when switching to an indirect context and back.
Presumably this bug will still happen if you do that while the context is
bound in another thread.

I don't think it should make any difference anyway if the flush is moved lower
in intelMakeCurrent because all of the code between the original flush and the
call to _mesa_make_current looks quite trivial and it only operates on the new
context not the old one.