dri: Change __DriverApiRec::CreateContext to take a struct for attribs
Previously the CreateContext method of __DriverApiRec took a set of arguments
to describe the attribute values from the window system API's
CreateContextAttribs function. As more attributes get added this could quickly
get unworkable and every new attribute needs a modification for every driver.

To fix that this patch passes the attribute values in a struct instead. The
struct has a bitmask to specify which members are used. However the first
three members (two for the GL version and one for the flags) are always set.
If the bit is not set in the attribute mask then it can be assumed the
attribute has the default value. That way the drivers can quickly check if any
attributes were set that they don't understand. When we add any new attributes
we can just extend the struct and add a new flag and the existing drivers will
automatically report an error when the attribute is used.

Currently the only attribute using this mechanism is the reset strategy but
the plan is to also use it for the flush control attribute in a later patch.