i965: Don't try to compact instructions with the EOT flag set
There is no flag for EOT in a compacted instruction. Normally this
doesn't matter because the final instructions are usually URB writes
which look like they have a strange immediate value for src1 and it
bails out when it can't compact it. However with the opt_sampler_eot
optimisation the sampler indirect messages end up with a register for
src1 so it will try to compact it but the EOT flag will be lost.

The following Piglit test hits this problem:

 spec@arb_gpu_shader5@execution@sampler_array_indexing@fs-simple

However this patch doesn't make the test pass because there is an
additional problem that the generator will try to use the destination
register as a temporary value but the destination will be null in this
case.

Cc: mesa-stable@lists.freedesktop.org