i965/sched: add a heuristic for scheduling uses of textures
When doing register pressure scheduling, we often have the case where a
texture is scheduled and then four instruction that together consume the
value it produces become available to schedule. The LIFO heuristic tried
to solve this by scheduling things as close to the parent that unblocked
them as possible, but we can do better. Instead, add a heuristic that
tries to schedule reads of registers where all the remaining reads are
ready to be scheduled. This helps us in most of the cases where LIFO
would have helped, without being so terrible for latency.