i965/sched: switch to register pressure scheduling dynamically
Now that we keep track of the register pressure, and we don't need to
fall back to the LIFO heuristic, it's trivial to be able to switch to a
different heuristic when our register pressure exceeds a certain
threshold, without worrying about which heuristic. We still need to try
to schedule multiple times, though, since for some shaders, a high
threshold will yield a better spill-free program than a low threshold,
while for other programs we might not be able to allocate without
spilling until we drop the threshold quite low. The parameters were
determined by increasing the tolerance until there weren't any more
SIMD16 programs that failed to compile because they spilled.