i965/skl: Set the extended function control bits to 0
On SKL+ the send and sendc messages have an extra large extended
message descriptor with an extra 16 bits called the extended function
control. Previously we were filling this with bits from the
description of src1 because the bits for it overlap with that. In
order to play it safe this patch makes it write 0 into it instead.