i965/fs: Eliminate some broadcast instructions
When a SHADER_OPCODE_FIND_LIVE_CHANNEL instruction is removed it will
now check if the following instruction is a broadcast instruction
using the result as a source. If so it will change it to a MOV
instruction instead. If the src and dst are the same it will
additionally just remove the broadcast entirely. This will likely
cause the find-live-channel instruction to be removed too via
dead-code elimination.

Ideally this would all happen automatically as a result of changing
the find-live-channel instruction to a mov instruction with a constant
source. However the constant doesn't end up getting propagated to the
move instruction because the mov is considered as a partial write.