i965/fs/skl+: Use ld_lz when the LOD parameter is zero
Adds an optimisation pass which recognises the LD sampler message type
when the LOD parameter is either a constant zero or not given and
replaces it with the LD_LZ message type. This is the same but the LOD
is hardcoded to zero and doesn't need to be in the message. This can
be a benefit for shaders using texelFetch with 3 coordinates because
otherwise the LOD parameter can't be optimised out because it comes
before the r coordinate.