i965/fs/gen9: Use SENDS for SHADER_OPCODE_TXF_CMS_W
The ld2dms_w instruction (which is used for SHADER_OPCODE_TXF_CMS_W)
is only available on Gen9. In that case the SENDS instruction is also
available which allows us to split the message payload into two
register regions. This is done for ld2dms_w in order to put the sample
index in the first region and the rest of the payload in the second.
The assumption is that any shader calling texelFetch on a sampler2DMS
is likely to fetch more than one sample using the same texture
coordinates. If the payload is split up like this then the register
allocator is able to resuse the register for the second part of the
payload for all calls to texelFetch. This is particularly useful the
16x MSAA resolve shader because in that case without this patch there
is two much register pressure for the SIMD16 version and it falls back
to SIMD8 only.