i965/fs/gen9: Use SENDS for TXF_MCS instructions
This adds a new pseudo opcode called TXF_MCS_SPLIT which is the same
as TXF_MCS except that the payload does not include the sample index
and it is instead split into a separate argument. This is later used
to make a SENDS instruction with the sample index in a separate
register region. In the likely case that the shader calls texelFetch
with different sample indices and identical texture coordinates this
will allow the register coalescer to combine the registers for the
texture coordinates. This new instruction is always used instead of
TXF_MCS on Gen9.

total instructions in shared programs: 7874519 -> 7874355 (-0.00%)
instructions in affected programs:     3022 -> 2858 (-5.43%)
total loops in shared programs:        2131 -> 2131 (0.00%)
helped:                                26

Note, I had to hack Mesa to default to assuming the samplers use CMS
layout in order to get these results because otherwise the precompiler
assumes it's UMS and doesn't use the ld2dms instruction. 8 of the
helped shaders are actually just copies of the meta resolve shader.