i965/meta-fast-clear: Convert the clear color through the surf format
When programming the fast clear color there was previously a chunk of
code to try to make the color match the constraints of the surface
format such as by filling in missing components and handling luminance
formats. These cases are not handled by the hardware. There are some
additional possible restrictions that the hardware does seem to
handle, such as clamping to [0,1] for normalised formats. However for
whatever reason it doesn't clamp to [0,∞] for the special float
formats that don't have a sign bit. Rather than adding yet another
special case for this format this patch makes it instead convert the
color to the actual surface format and back again so that we can be
sure it will have all of the possible restrictions. Additionally this
would avoid some other potential surprises such as getting more
precision for the clear color when fast clears are used.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=93338