mesa: Add the GL_MESA_texture_view_rgb_32 extension
If this extension is advertised then the upper layers of Mesa will
allow a texure view to be created between any formats within the
GL_VIEW_CLASS_24 and GL_VIEW_CLASS_32 classes. The expectation is that
drivers that implement 24-bit RGB textures by padding them out to 32
bits with an unused X component will trivially implement this
extension.