intel/compiler: Use emit_emplace in the builders
Replaces all calls of the form emit(instruction(…)) with
emit_emplace(…). This should avoid a redundant copy of the
stack-allocated temporary instruction. Although the temporary would
have been stack-allocated so it might not seem like a big deal, the
instructions do also have an internal allocation for the sources so we
also get to avoid an extra allocation.

I tested this with shader-db over four iterations and the average sum
of the time reported for all of the threads drops from 406 seconds to
366.