compiler: Add new system value SYSTEM_VALUE_BASE_VERTEX_ID
This VS system value will contain the value passed as <basevertex>
for indexed draw calls or the value passed as <first> for non-indexed
draw calls. It will be used to calculate the gl_VertexID as
SYSTEM_VALUE_VERTEX_ID_ZERO_BASE plus SYSTEM_VALUE_BASE_VERTEX_ID.
Note that the current calculation which uses SYSTEM_VALUE_BASE_VERTEX
is not right, as gl_BaseVertex should be zero for non-indexed calls.

Reviewed-by: Neil Roberts <nroberts@igalia.com>