st/glsl_to_tgsi: Add support for SYSTEM_VALUE_BASE_VERTEX_ID
SYSTEM_VALUE_BASE_VERTEX has changed to be the correct value for
gl_BaseVertexID, which means it will be zero when used with a
non-indexed call. The new BASE_VERTEX_ID value can be used as before
as an offset to calculate a value for gl_VertexID. These values should
be different, but this patch just makes them same for now in order to
at least retain the previous behaviour and not break gl_BaseVertexID
and gl_VertexID entirely on radeonsi.